requirements.txt
Django==4.2.7
python-docx==0.8.11
PyPDF2==3.0.1
openai==1.3.5
python-decouple==3.8
Pillow==10.1.0
gunicorn==21.2.0